Fixes relevant when building the CSL version on a 32-bit system - or to be more precise on a system that does not support a 128-bit integer data-type.
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5004 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/arith04.cpp
@@ -586,7 +586,7 @@ static LispObject rationalizef128(float128_t *dd)
         v0 = v1; v1 = v2;
     }
     LispObject p1;
-    if (f128M_negative(dd)) p1 = make_lisp_integer128(-int128(u1));
+    if (f128M_negative(dd)) p1 = make_lisp_integer128(-u1);
     else p1 = make_lisp_unsigned128(u1);
     if (v1 == 1) return p1;
     push(p1);
